+++ psychopy/app/builder/components/_visual.py
@@ -18,6 +18,7 @@ class VisualComponent(_base.BaseComponent):
             updates='constant', allowedUpdates=['constant','set every repeat','set every frame'],
             hint="Colour of this stimulus (e.g. [1,1,0], 'red' )")
         self.params['colourSpace']=Param(colourSpace, valType='code', allowedVals=['rgb','dkl','lms'],
+            updates='constant', allowedUpdates=['constant'],
             hint="Choice of colour space for the colour (rgb, dkl, lms)")
         self.params['pos']=Param(pos, valType='code', allowedTypes=[],
             updates='constant', allowedUpdates=['constant','set every repeat','set every frame'],
@@ -34,12 +35,26 @@ class VisualComponent(_base.BaseComponent):
             
     def writeFrameCode(self,buff):
         """Write the code that will be called every frame
-        """    
+        """
         self.writeTimeTestCode(buff)#writes an if statement to determine whether to draw etc
         buff.setIndentLevel(1, relative=True)#because of the 'if' statement of the times test
         #set parameters that need updating every frame
-        self.writeParamUpdates(buff, 'frame')
+        self.writeParamUpdates(buff, 'set every frame')
         #draw the stimulus
         buff.writeIndented("%(name)s.draw()\n" %(self.params))
         buff.setIndentLevel(-1, relative=True)
-
+       
+    def writeParamUpdates(self, buff, updateType):
+        """write updates to the buffer for each parameter that needs it
+        updateType can be 'experiment', 'routine' or 'frame'
+        """
+        for thisParamName in self.params.keys():
+            thisParam=self.params[thisParamName]
+            if thisParamName=='colour':
+                #we need setRGB=colour (not setColour=colour)
+                thisParamName= self.params['colourSpace'].val.upper()#thisParam is the correct value, but the name is the space!
+            else:
+                thisParamName = thisParamName.capitalize()
+            if thisParam.updates==updateType:
+                buff.writeIndented("%s.set%s(%s)\n" %(self.params['name'], thisParamName, thisParam)) 
+